requirements.txt
open_clip_torch
faiss-cpu
gradio>=4.0
torch>=2.0
torchvision
numpy
pandas
Pillow
pycocotools
tqdm
requests
